requirements.txt
Flask
Flask-SQLAlchemy
pymysql
Flask-migrate
cryptography
python-dotenv
